backend/requirements.txt
Django>=5.0
mysqlclient>=2.2
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
drf-spectacular>=0.27
python-decouple>=3.8
celery>=5.3
redis>=5.0
django-redis>=5.4
Pillow>=10.3
django-filter>=24.0
phonenumbers>=8.13
requests>=2.31
